+++ docs/data/base/components/trap-focus/BasicTrapFocus.tsx.preview
