+++ docs/data/base/components/trap-focus/LazyTrapFocus.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 import Box from '@mui/material/Box';
-import Portal from '@mui/material/Portal';
-import TrapFocus from '@mui/material/Unstable_TrapFocus';
+import TrapFocus from '@mui/base/TrapFocus';
 
-export default function PortalTrapFocus() {
+export default function LazyTrapFocus() {
   const [open, setOpen] = React.useState(false);
-  const [container, setContainer] = React.useState(null);
 
   return (
     <Box
@@ -19,26 +17,18 @@ export default function PortalTrapFocus() {
         Open
       </button>
       {open && (
-        <TrapFocus open>
+        <TrapFocus open disableAutoFocus>
           <Box tabIndex={-1} sx={{ mt: 1, p: 1 }}>
             <label>
               First name: <input type="text" />
             </label>
             <br />
-            <Portal container={container}>
-              <label>
-                Last name: <input type="text" />
-              </label>
-              <br />
-            </Portal>
             <button type="button" onClick={() => setOpen(false)}>
               Close
             </button>
           </Box>
         </TrapFocus>
       )}
-
-      <div ref={setContainer} />
     </Box>
   );
 }
