+++ docs/data/base/components/trap-focus/LazyTrapFocus.tsx.preview
