+++ docs/data/base/components/trap-focus/PortalTrapFocus.js
@@ -1,9 +1,11 @@
 import * as React from 'react';
-import Box from '@mui/material/Box';
-import TrapFocus from '@mui/material/Unstable_TrapFocus';
+import { Box } from '@mui/system';
+import Portal from '@mui/base/Portal';
+import TrapFocus from '@mui/base/TrapFocus';
 
-export default function DisableEnforceFocus() {
+export default function PortalTrapFocus() {
   const [open, setOpen] = React.useState(false);
+  const [container, setContainer] = React.useState(null);
 
   return (
     <Box
@@ -17,18 +19,26 @@ export default function DisableEnforceFocus() {
         Open
       </button>
       {open && (
-        <TrapFocus disableEnforceFocus open>
+        <TrapFocus open>
           <Box tabIndex={-1} sx={{ mt: 1, p: 1 }}>
             <label>
               First name: <input type="text" />
             </label>
             <br />
+            <Portal container={container}>
+              <label>
+                Last name: <input type="text" />
+              </label>
+              <br />
+            </Portal>
             <button type="button" onClick={() => setOpen(false)}>
               Close
             </button>
           </Box>
         </TrapFocus>
       )}
+
+      <div ref={setContainer} />
     </Box>
   );
 }
