+++ docs/pages/base/api/trap-focus.json
@@ -11,11 +11,11 @@
       "default": "function defaultIsEnabled() {\n  return true;\n}"
     }
   },
-  "name": "Unstable_TrapFocus",
+  "name": "TrapFocus",
   "styles": { "classes": [], "globalClasses": {}, "name": null },
   "spread": false,
-  "filename": "/packages/mui-base/src/Unstable_TrapFocus/Unstable_TrapFocus.js",
+  "filename": "/packages/mui-base/src/TrapFocus/TrapFocus.js",
   "inheritance": null,
-  "demos": "<ul><li><a href=\"/components/trap-focus/\">Trap Focus</a></li></ul>",
+  "demos": "<ul><li><a href=\"/base/components/trap-focus/\">Trap Focus</a></li></ul>",
   "cssComponent": false
 }
