--- docs/pages/base/api/unstable-trap-focus.json
@@ -1,21 +1,0 @@
-{
-  "props": {
-    "open": { "type": { "name": "bool" }, "required": true },
-    "children": { "type": { "name": "custom", "description": "element" } },
-    "disableAutoFocus": { "type": { "name": "bool" } },
-    "disableEnforceFocus": { "type": { "name": "bool" } },
-    "disableRestoreFocus": { "type": { "name": "bool" } },
-    "getTabbable": { "type": { "name": "func" } },
-    "isEnabled": {
-      "type": { "name": "func" },
-      "default": "function defaultIsEnabled() {\n  return true;\n}"
-    }
-  },
-  "name": "Unstable_TrapFocus",
-  "styles": { "classes": [], "globalClasses": {}, "name": null },
-  "spread": false,
-  "filename": "/packages/mui-base/src/Unstable_TrapFocus/Unstable_TrapFocus.js",
-  "inheritance": null,
-  "demos": "<ul><li><a href=\"/material-ui/react-trap-focus/\">Trap Focus</a></li></ul>",
-  "cssComponent": false
-}
