--- docs/pages/material-ui/react-trap-focus.js
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import {
-  demos,
-  docs,
-  demoComponents,
-} from 'docs/data/material/components/trap-focus/trap-focus.md?@mui/markdown';
-
-export default function Page() {
-  return <MarkdownDocs demos={demos} docs={docs} demoComponents={demoComponents} />;
-}
