+++ docs/src/modules/utils/replaceMarkdownLinks.test.js
@@ -198,7 +198,7 @@ describe('replaceMarkdownLinks', () => {
       [ButtonUnstyled](/api/button-unstyled)
       [TabPanelUnstyled](/api/tab-panel-unstyled)
       [TabsListUnstyled](/api/tabs-list-unstyled)
-      [Unstable_TrapFocus](/api/unstable-trap-focus)
+      [TrapFocus](/api/trap-focus)
       [ClickAwayListener](/api/click-away-listener)
       [IconButton](/api/icon-button)
       [LoadingButton](/api/loading-button)
@@ -217,7 +217,7 @@ describe('replaceMarkdownLinks', () => {
       [ButtonUnstyled](/base/api/button-unstyled)
       [TabPanelUnstyled](/base/api/tab-panel-unstyled)
       [TabsListUnstyled](/base/api/tabs-list-unstyled)
-      [Unstable_TrapFocus](/base/api/unstable-trap-focus)
+      [TrapFocus](/base/api/trap-focus)
       [ClickAwayListener](/base/api/click-away-listener)
       [IconButton](/material-ui/api/icon-button)
       [LoadingButton](/material-ui/api/loading-button)
