+++ docs/translations/api-docs/trap-focus/trap-focus-pt.json
