+++ packages/mui-base/src/TrapFocus/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './TrapFocus';
+export * from './TrapFocus';
