+++ packages/mui-base/src/TrapFocus/index.js
@@ -1,0 +1,1 @@
+export { default } from './TrapFocus';
