--- packages/mui-base/src/Unstable_TrapFocus/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './Unstable_TrapFocus';
-export * from './Unstable_TrapFocus';
