--- packages/mui-base/src/Unstable_TrapFocus/index.js
@@ -1,1 +1,0 @@
-export { default } from './Unstable_TrapFocus';
