+++ packages/mui-base/src/index.d.ts
@@ -78,5 +78,5 @@ export * from './TabUnstyled';
 export { default as TextareaAutosize } from './TextareaAutosize';
 export * from './TextareaAutosize';
 
-export { default as Unstable_TrapFocus } from './Unstable_TrapFocus';
-export * from './Unstable_TrapFocus';
+export { default as TrapFocus } from './TrapFocus';
+export * from './TrapFocus';
