+++ packages/mui-base/src/index.js
@@ -72,4 +72,4 @@ export * from './TabUnstyled';
 
 export { default as TextareaAutosize } from './TextareaAutosize';
 
-export { default as Unstable_TrapFocus } from './Unstable_TrapFocus';
+export { default as TrapFocus } from './TrapFocus';
