+++ packages/mui-material/src/Unstable_TrapFocus/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from '@mui/base/Unstable_TrapFocus';
-export * from '@mui/base/Unstable_TrapFocus';
+export { default } from '@mui/base/TrapFocus';
+export * from '@mui/base/TrapFocus';
